+++ src/app/components/tables/data-table/interfaces/terra-data-table-row.interface.ts
@@ -1,5 +1,4 @@
 import { TerraDataTableCellInterface } from './terra-data-table-cell.interface';
-import { TerraDataTableContextMenuEntryInterface } from '../context-menu/data/terra-data-table-context-menu-entry.interface';
 
 /**
  * @author mkunze
@@ -9,7 +8,6 @@ export interface TerraDataTableRowInterface<D>
     cellList:Array<TerraDataTableCellInterface>;
     data?:D;
     clickFunction?:() => void;
-    contextMenuLinkList?:Array<TerraDataTableContextMenuEntryInterface<D>>;
     isActive?:boolean;
     disabled?:boolean;
     selected?:boolean;